+++ src/main/java/in/koreatech/koin/domain/bus/batch/client/TmoneyExpressBusClient.java
@@ -1,9 +1,9 @@
-package in.koreatech.koin.domain.bus.client;
+package in.koreatech.koin.domain.bus.batch.client;
 
 import static java.net.URLEncoder.encode;
+import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.time.format.DateTimeFormatter.ofPattern;
 
-import java.net.URL;
 import java.time.Clock;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
@@ -21,15 +21,17 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
+import org.springframework.web.util.UriComponents;
+import org.springframework.web.util.UriComponentsBuilder;
 
-import in.koreatech.koin.domain.bus.dto.express.PublicOpenApiResponse;
-import in.koreatech.koin.domain.bus.exception.BusOpenApiException;
+import in.koreatech.koin.domain.bus.batch.response.TmoneyOpenApiResponse;
+import in.koreatech.koin.domain.bus.batch.exception.BusOpenApiException;
 import in.koreatech.koin.domain.bus.model.enums.BusStation;
 import in.koreatech.koin.domain.bus.model.enums.ExpressBusStationNode;
 import in.koreatech.koin.domain.bus.model.express.ExpressBusCache;
 import in.koreatech.koin.domain.bus.model.express.ExpressBusCacheInfo;
 import in.koreatech.koin.domain.bus.model.express.ExpressBusRoute;
-import in.koreatech.koin.domain.bus.model.express.PublicOpenApiExpressBusArrival;
+import in.koreatech.koin.domain.bus.batch.response.TmoneyOpenApiExpressBusArrival;
 import in.koreatech.koin.domain.bus.repository.ExpressBusCacheRepository;
 import in.koreatech.koin.domain.version.model.VersionType;
 import in.koreatech.koin.domain.version.repository.VersionRepository;
@@ -38,21 +40,20 @@ import in.koreatech.koin.global.exception.KoinIllegalStateException;
 import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
 
 /**
- * OpenApi 상세: 국토교통부_(TAGO)_버스도착정보
- * https://www.data.go.kr/tcs/dss/selectApiDataDetailView.do?publicDataPk=15098541
+ * OpenApi 상세: 티머니 Api
+ * https://apiportal.tmoney.co.kr:18443/apiGallery/apiGalleryDetail.do?apiId=API201906241410183kp&apiPckgId=APK2024051316462950w&isTestYn=Y
  */
 @Component
-@CallControlInfo(ratio = 2)
-public class PublicExpressBusClient extends ExpressBusClient {
+@CallControlInfo(ratio = 9)
+public class TmoneyExpressBusClient extends ExpressBusClient {
 
-    private static final String OPEN_API_URL = "https://apis.data.go.kr/1613000/SuburbsBusInfoService/getStrtpntAlocFndSuberbsBusInfo";
-    private static final String ENCODE_TYPE = "UTF-8";
+    private static final String OPEN_API_URL = "https://apigw.tmoney.co.kr:5556/gateway/xzzIbtListGet/v1/ibt_list";
 
     private final RestTemplate restTemplate;
     private final String openApiKey;
 
-    public PublicExpressBusClient(
-        @Value("${OPEN_API_KEY_PUBLIC}") String openApiKey,
+    public TmoneyExpressBusClient(
+        @Value("${OPEN_API_KEY_TMONEY}") String openApiKey,
         VersionRepository versionRepository,
         RestTemplate restTemplate,
         Clock clock,
@@ -65,35 +66,29 @@ public class PublicExpressBusClient extends ExpressBusClient {
 
     @Override
     @Transactional(propagation = Propagation.REQUIRES_NEW)
-    @CircuitBreaker(name = "publicExpressBus")
+    @CircuitBreaker(name = "tmoneyExpressBus")
     public void storeRemainTime() {
         for (BusStation depart : BusStation.values()) {
             for (BusStation arrival : BusStation.values()) {
                 if (depart == arrival || depart.equals(BusStation.STATION) || arrival.equals(BusStation.STATION)) {
                     continue;
                 }
-                PublicOpenApiResponse openApiResponse;
+                TmoneyOpenApiResponse openApiResponse;
                 try {
                     openApiResponse = getOpenApiResponse(depart, arrival);
                 } catch (BusOpenApiException e) {
                     continue;
                 }
-                List<PublicOpenApiExpressBusArrival> busArrivals = extractBusArrivalInfo(openApiResponse);
+                List<TmoneyOpenApiExpressBusArrival> busArrivals = extractBusArrivalInfo(openApiResponse);
 
                 ExpressBusCache expressBusCache = ExpressBusCache.of(
                     new ExpressBusRoute(depart.getName(), arrival.getName()),
-                    // API로 받은 yyyyMMddHHmm 형태의 시간을 HH:mm 형태로 변환하여 Redis에 저장한다.
+                    // API로 받은 HHmm 형태의 시간을 HH:mm 형태로 변환하여 Redis에 저장한다.
                     busArrivals.stream()
                         .map(it -> new ExpressBusCacheInfo(
-                            LocalTime.parse(
-                                LocalDateTime.parse(it.depPlandTime(), ofPattern("yyyyMMddHHmm"))
-                                    .format(ofPattern("HH:mm"))
-                            ),
-                            LocalTime.parse(
-                                LocalDateTime.parse(it.arrPlandTime(), ofPattern("yyyyMMddHHmm"))
-                                    .format(ofPattern("HH:mm"))
-                            ),
-                            it.charge()
+                            LocalTime.parse(it.TIM_TIM_O(), ofPattern("HHmm")),
+                            LocalTime.parse(it.TIM_TIM_O(), ofPattern("HHmm")).plusMinutes(it.LIN_TIM()),
+                            1900
                         ))
                         .toList()
                 );
@@ -105,19 +100,27 @@ public class PublicExpressBusClient extends ExpressBusClient {
         versionRepository.getByType(VersionType.EXPRESS).update(clock);
     }
 
-    private PublicOpenApiResponse getOpenApiResponse(BusStation depart, BusStation arrival) {
+    private List<TmoneyOpenApiExpressBusArrival> extractBusArrivalInfo(TmoneyOpenApiResponse tmoneyResponse) {
+        if (!tmoneyResponse.code().equals("00000")) {
+            return Collections.emptyList();
+        }
+        return tmoneyResponse.response().LINE_LIST();
+    }
+
+    private TmoneyOpenApiResponse getOpenApiResponse(BusStation depart, BusStation arrival) {
         try {
             HttpHeaders headers = new HttpHeaders();
             headers.setContentType(MediaType.APPLICATION_JSON);
+            headers.set("x-Gateway-APIKey", openApiKey);
             headers.set("Accept", "*/*");
 
             HttpEntity<?> entity = new HttpEntity<>(headers);
-            URL uri = new URL(getRequestURL(depart, arrival));
-            ResponseEntity<PublicOpenApiResponse> response = restTemplate.exchange(
-                uri.toURI(),
+            UriComponents uri = getRequestURL(depart, arrival);
+            ResponseEntity<TmoneyOpenApiResponse> response = restTemplate.exchange(
+                uri.toString(),
                 HttpMethod.GET,
                 entity,
-                PublicOpenApiResponse.class
+                TmoneyOpenApiResponse.class
             );
             return response.getBody();
         } catch (HttpClientErrorException e) {
@@ -128,31 +131,24 @@ public class PublicExpressBusClient extends ExpressBusClient {
         }
     }
 
-    private String getRequestURL(BusStation depart, BusStation arrival) {
+    private UriComponents getRequestURL(BusStation depart, BusStation arrival) {
         ExpressBusStationNode departNode = ExpressBusStationNode.from(depart);
         ExpressBusStationNode arrivalNode = ExpressBusStationNode.from(arrival);
-        StringBuilder urlBuilder = new StringBuilder(OPEN_API_URL); /*URL*/
+        LocalDateTime today = LocalDateTime.now(clock);
+        UriComponents uri;
         try {
-            urlBuilder.append("?" + encode("serviceKey", ENCODE_TYPE) + "=" + encode(openApiKey, ENCODE_TYPE));
-            urlBuilder.append("&" + encode("numOfRows", ENCODE_TYPE) + "=" + encode("30", ENCODE_TYPE));
-            urlBuilder.append("&" + encode("_type", ENCODE_TYPE) + "=" + encode("json", ENCODE_TYPE));
-            urlBuilder.append(
-                "&" + encode("depTerminalId", ENCODE_TYPE) + "=" + encode(departNode.getStationId(), ENCODE_TYPE));
-            urlBuilder.append(
-                "&" + encode("arrTerminalId", ENCODE_TYPE) + "=" + encode(arrivalNode.getStationId(), ENCODE_TYPE));
-            urlBuilder.append("&" + encode("depPlandTime", ENCODE_TYPE) + "="
-                + encode(LocalDateTime.now(clock).format(ofPattern("yyyyMMdd")), ENCODE_TYPE));
-            return urlBuilder.toString();
+            uri = UriComponentsBuilder
+                .fromHttpUrl(OPEN_API_URL)
+                .pathSegment(encode(today.format(ofPattern("yyyyMMdd")), UTF_8))
+                .pathSegment("0000")
+                .pathSegment(encode(departNode.getTmoneyStationId(), UTF_8))
+                .pathSegment(encode(arrivalNode.getTmoneyStationId(), UTF_8))
+                .pathSegment("9")
+                .pathSegment("0")
+                .build();
+            return uri;
         } catch (Exception e) {
-            throw new KoinIllegalStateException("공공데이터포털 시외버스 API URL 생성중 문제가 발생했습니다.", "uri:" + urlBuilder);
+            throw new KoinIllegalStateException("티머니 시외버스 API URL 생성중 문제가 발생했습니다. uri: " + OPEN_API_URL);
         }
-    }
-
-    private List<PublicOpenApiExpressBusArrival> extractBusArrivalInfo(PublicOpenApiResponse publicResponse) {
-        if (!publicResponse.response().header().resultCode().equals("00")
-            || publicResponse.response().body().totalCount() == 0) {
-            return Collections.emptyList();
-        }
-        return publicResponse.response().body().items().item();
     }
 }
