+++ src/main/java/in/koreatech/koin/domain/bus/batch/util/BusScheduler.java
@@ -1,10 +1,10 @@
-package in.koreatech.koin.domain.bus.util;
+package in.koreatech.koin.domain.bus.batch.util;
 
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
-import in.koreatech.koin.domain.bus.service.CityBusService;
-import in.koreatech.koin.domain.bus.service.ExpressBusService;
+import in.koreatech.koin.domain.bus.batch.service.CityBusCacheService;
+import in.koreatech.koin.domain.bus.batch.service.ExpressBusCacheService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -13,14 +13,14 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class BusScheduler {
 
-    private final ExpressBusService expressBusService;
-    private final CityBusService cityBusService;
+    private final CityBusCacheService cityBusCacheService;
+    private final ExpressBusCacheService expressBusCacheService;
 
     @Scheduled(cron = "0 * * * * *")
     public void cacheCityBusByOpenApi() {
         try {
-            cityBusService.cacheRemainTime();
-            cityBusService.cacheCityBusRoute();
+            cityBusCacheService.cacheRemainTime();
+            cityBusCacheService.cacheCityBusRoute();
         } catch (Exception e) {
             log.warn("시내버스 스케줄링 과정에서 오류가 발생했습니다.");
         }
@@ -29,7 +29,7 @@ public class BusScheduler {
     @Scheduled(cron = "0 30 0 * * *")
     public void cacheExpressBusByOpenApi() {
         try {
-            expressBusService.cacheRemainTimeByRatio();
+            expressBusCacheService.cacheRemainTimeByRatio();
         } catch (Exception e) {
             log.warn("시외버스 스케줄링 과정에서 오류가 발생했습니다.");
         }
