+++ src/main/java/in/koreatech/koin/domain/bus/service/CityBusService.java
@@ -2,10 +2,7 @@ package in.koreatech.koin.domain.bus.service;
 
 import static in.koreatech.koin.domain.bus.model.enums.BusStation.getDirection;
 
-import java.time.Clock;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
@@ -15,27 +12,17 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import in.koreatech.koin.domain.bus.client.CityBusClient;
-import in.koreatech.koin.domain.bus.client.CityBusRouteClient;
 import in.koreatech.koin.domain.bus.dto.BusScheduleResponse;
-import in.koreatech.koin.domain.bus.dto.city.CityBusArrival;
-import in.koreatech.koin.domain.bus.dto.city.CityBusRoute;
 import in.koreatech.koin.domain.bus.dto.city.CityBusTimetableResponse;
-import in.koreatech.koin.domain.bus.exception.BusOpenApiException;
-import in.koreatech.koin.domain.bus.model.city.CityBusCache;
-import in.koreatech.koin.domain.bus.model.city.CityBusCacheInfo;
 import in.koreatech.koin.domain.bus.model.city.CityBusRemainTime;
 import in.koreatech.koin.domain.bus.model.city.CityBusTimetable;
 import in.koreatech.koin.domain.bus.model.enums.BusDirection;
 import in.koreatech.koin.domain.bus.model.enums.BusStation;
-import in.koreatech.koin.domain.bus.model.enums.BusStationNode;
 import in.koreatech.koin.domain.bus.model.enums.CityBusDirection;
 import in.koreatech.koin.domain.bus.repository.CityBusCacheRepository;
 import in.koreatech.koin.domain.bus.repository.CityBusRouteCacheRepository;
 import in.koreatech.koin.domain.bus.repository.CityBusTimetableRepository;
 import in.koreatech.koin.domain.bus.service.route.CityBusRouteManager;
-import in.koreatech.koin.domain.version.model.VersionType;
-import in.koreatech.koin.domain.version.repository.VersionRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -48,11 +35,6 @@ public class CityBusService {
     private final CityBusRouteCacheRepository cityBusRouteCacheRepository;
     private final CityBusCacheRepository cityBusCacheRepository;
     private final CityBusTimetableRepository cityBusTimetableRepository;
-    private final CityBusClient cityBusClient;
-    private final CityBusRouteClient cityBusRouteClient;
-
-    private final VersionRepository versionRepository;
-    private final Clock clock;
 
     public List<CityBusRemainTime> getBusRemainTime(BusStation depart, BusStation arrival) {
         BusDirection direction = getDirection(depart, arrival);
@@ -93,42 +75,5 @@ public class CityBusService {
         CityBusTimetable timetable = cityBusTimetableRepository
             .getByBusInfoNumberAndBusInfoArrival(busNumber, arrival.getName());
         return CityBusRouteManager.getCityBusSchedule(timetable, busNumber, depart, arrival, date);
-    }
-
-    @Transactional
-    public void cacheRemainTime() {
-        LocalDateTime updatedAt = LocalDateTime.now(clock);
-        BusStationNode.getNodeIds().stream()
-            .map(nodeId -> {
-                try {
-                    return cityBusClient.fetchBusArrivalList(nodeId);
-                } catch (BusOpenApiException ignored) {
-                    return Collections.<CityBusArrival>emptyList();
-                }
-            })
-            .filter(arrivalInfos -> !arrivalInfos.isEmpty())
-            .map(arrivalInfos -> CityBusCache.of(
-                arrivalInfos.get(0).nodeid(),
-                toCityBusCacheInfo(arrivalInfos, updatedAt)
-            ))
-            .forEach(cityBusCacheRepository::save);
-        versionRepository.getByType(VersionType.CITY).update(clock);
-    }
-
-    private List<CityBusCacheInfo> toCityBusCacheInfo(List<CityBusArrival> arrivalInfos, LocalDateTime updatedAt) {
-        return arrivalInfos.stream()
-            .map(busArrivalInfo -> busArrivalInfo.toCityBusCacheInfo(updatedAt))
-            .toList();
-    }
-
-    @Transactional
-    public void cacheCityBusRoute() {
-        BusStationNode.getNodeIds().forEach((nodeId) -> {
-            try {
-                Set<CityBusRoute> routes = Set.copyOf(cityBusRouteClient.fetchBusRouteList(nodeId));
-                cityBusRouteCacheRepository.save(CityBusRoute.toCityBusRouteCache(nodeId, routes));
-            } catch (BusOpenApiException ignored) {
-            }
-        });
     }
 }
