+++ src/test/java/in/koreatech/koin/AcceptanceTest.java
@@ -22,9 +22,9 @@ import org.testcontainers.utility.DockerImageName;
 import in.koreatech.koin.config.TestJpaConfiguration;
 import in.koreatech.koin.config.TestRedisConfiguration;
 import in.koreatech.koin.config.TestTimeConfig;
-import in.koreatech.koin.domain.bus.client.PublicExpressBusClient;
-import in.koreatech.koin.domain.bus.client.StaticExpressBusClient;
-import in.koreatech.koin.domain.bus.client.TmoneyExpressBusClient;
+import in.koreatech.koin.domain.bus.batch.client.PublicExpressBusClient;
+import in.koreatech.koin.domain.bus.batch.client.StaticExpressBusClient;
+import in.koreatech.koin.domain.bus.batch.client.TmoneyExpressBusClient;
 import in.koreatech.koin.domain.community.keyword.model.ArticleKeywordEventListener;
 import in.koreatech.koin.domain.coop.model.CoopEventListener;
 import in.koreatech.koin.domain.owner.model.OwnerEventListener;
